+++ src/classification/index.ts
@@ -1,3 +1,4 @@
 export * from './computed';
+export * from './hmr.class';
 export * from './property';
 export * from './vue.class';